# lib/api/openapi.yaml
openapi: "3.1.0"
info:
  title: "HAY2010 Stock API"
  version: "1.0.0"
  description: "API REST pour la gestion des partenaires, documents, stock et affaires"
servers:
  - url: "/api/v1"
    description: "Base path for API v1"
security:
  - ApiKeyAuth: []
paths:
...
  /produits:
    get:
      operationId: listProducts
      summary: "Lister les produits"
      description: "Retourne une liste paginée de produits avec filtrage optionnel par catégorie, famille, statut actif, recherche textuelle, et tri."
      tags:
        - Produits
      security:
        - ApiKeyAuth: []
      parameters:
        - name: categorie
          in: query
          schema:
            type: integer
          description: "Filtrer par ID de catégorie"
        - name: famille
          in: query
          schema:
            type: string
          description: "Filtrer par famille de produit"
        - name: actif
          in: query
          schema:
            type: string
            enum: ["true", "false"]
          description: "Filtrer par statut actif"
        - name: search
          in: query
          schema:
            type: string
          description: "Recherche dans nom, code et code-barre EAN"
        - name: sort
          in: query
          schema:
            type: string
            enum: [id_produit, code_produit, nom_produit, famille, prix_vente, prix_achat, date_creation, date_modification]
            default: nom_produit
          description: "Champ de tri"
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
          description: "Ordre de tri"
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: "Liste paginée des produits"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPaginatedResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "400":
          $ref: "#/components/responses/BadRequest"

    post:
      operationId: createProduct
      summary: "Créer un produit"
      description: "Crée un nouveau produit. Le champ cree_par est automatiquement renseigné."
      tags:
        - Produits
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreate"
      responses:
        "201":
          description: "Produit créé"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: "Code produit déjà utilisé"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /produits/{id}:
    get:
      operationId: getProductById
      summary: "Obtenir un produit par ID"
      tags:
        - Produits
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "Détails du produit"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      operationId: updateProduct
      summary: "Mettre à jour un produit"
      tags:
        - Produits
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: "Produit mis à jour"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: "Code produit déjà utilisé"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteProduct
      summary: "Désactiver un produit (soft delete)"
      description: "Désactive le produit (est_actif = false)."
      tags:
        - Produits
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: "Produit désactivé"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /categories-produits/{id}/produits:
    get:
      operationId: getCategoryProducts
      summary: "Lister les produits d'une catégorie"
      tags:
        - CategoriesProduits
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: "Liste paginée des produits"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPaginatedResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: "API key au format hay2010_sk_live_xxxxxxxxxxxxxxxx"
  responses:
    Unauthorized:
      description: "Clé API invalide ou manquante"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: "Requête invalide"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: "Ressource non trouvée"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasMore:
          type: boolean
    Product:
      type: object
      properties:
        id_produit:
          type: integer
        code_produit:
          type: string
        nom_produit:
          type: string
        id_categorie:
          type: integer
          nullable: true
        famille:
          type: string
          nullable: true
        description_produit:
          type: string
          nullable: true
        code_barre_ean:
          type: string
          nullable: true
        unite_mesure:
          type: string
        poids_kg:
          type: number
          nullable: true
        volume_m3:
          type: number
          nullable: true
        prix_achat:
          type: number
          nullable: true
        prix_dernier_achat:
          type: number
          nullable: true
        coefficient:
          type: number
        prix_vente:
          type: number
          nullable: true
        prix_gros:
          type: number
          nullable: true
        taux_tva:
          type: number
        type_suivi_stock:
          type: string
        quantite_min_commande:
          type: integer
        niveau_reappro_quantite:
          type: integer
        stock_minimum:
          type: integer
        stock_maximum:
          type: integer
          nullable: true
        activer_suivi_stock:
          type: boolean
        id_fournisseur_principal:
          type: integer
          nullable: true
        reference_fournisseur:
          type: string
          nullable: true
        delai_livraison_fournisseur_jours:
          type: integer
          nullable: true
        est_actif:
          type: boolean
        en_sommeil:
          type: boolean
        est_abandonne:
          type: boolean
        date_creation:
          type: string
          format: date-time
        date_modification:
          type: string
          format: date-time
        cree_par:
          type: string
          nullable: true
        modifie_par:
          type: string
          nullable: true
        compte_general_vente:
          type: string
          nullable: true
        compte_general_achat:
          type: string
          nullable: true
        code_taxe_vente:
          type: string
          nullable: true
        code_taxe_achat:
          type: string
          nullable: true
    ProductCreate:
      type: object
      required:
        - code_produit
        - nom_produit
      properties:
        code_produit:
          type: string
          minLength: 1
          maxLength: 100
        nom_produit:
          type: string
          minLength: 1
          maxLength: 500
        id_categorie:
          type: integer
          nullable: true
        famille:
          type: string
          maxLength: 100
          nullable: true
        description_produit:
          type: string
          nullable: true
        code_barre_ean:
          type: string
          maxLength: 50
          nullable: true
        unite_mesure:
          type: string
          maxLength: 20
          default: U
        poids_kg:
          type: number
          nullable: true
        volume_m3:
          type: number
          nullable: true
        prix_achat:
          type: number
          nullable: true
        prix_dernier_achat:
          type: number
          nullable: true
        coefficient:
          type: number
          default: 1.0
        prix_vente:
          type: number
          nullable: true
        prix_gros:
          type: number
          nullable: true
        taux_tva:
          type: number
          minimum: 0
          maximum: 100
          default: 20.00
        type_suivi_stock:
          type: string
          maxLength: 20
          default: AUCUN
        quantite_min_commande:
          type: integer
          default: 1
        niveau_reappro_quantite:
          type: integer
          default: 0
        stock_minimum:
          type: integer
          default: 0
        stock_maximum:
          type: integer
          nullable: true
        activer_suivi_stock:
          type: boolean
        id_fournisseur_principal:
          type: integer
          nullable: true
        reference_fournisseur:
          type: string
          maxLength: 100
          nullable: true
        delai_livraison_fournisseur_jours:
          type: integer
          nullable: true
        compte_general_vente:
          type: string
          maxLength: 20
          nullable: true
        compte_general_achat:
          type: string
          maxLength: 20
          nullable: true
        code_taxe_vente:
          type: string
          maxLength: 10
          nullable: true
        code_taxe_achat:
          type: string
          maxLength: 10
          nullable: true
    ProductUpdate:
      type: object
      description: "Tous les champs sont optionnels pour la mise à jour partielle"
      properties:
        code_produit:
          type: string
          minLength: 1
          maxLength: 100
        nom_produit:
          type: string
          minLength: 1
          maxLength: 500
        id_categorie:
          type: integer
          nullable: true
        famille:
          type: string
          maxLength: 100
          nullable: true
        description_produit:
          type: string
          nullable: true
        code_barre_ean:
          type: string
          maxLength: 50
          nullable: true
        unite_mesure:
          type: string
          maxLength: 20
        poids_kg:
          type: number
          nullable: true
        volume_m3:
          type: number
          nullable: true
        prix_achat:
          type: number
          nullable: true
        prix_dernier_achat:
          type: number
          nullable: true
        coefficient:
          type: number
        prix_vente:
          type: number
          nullable: true
        prix_gros:
          type: number
          nullable: true
        taux_tva:
          type: number
          minimum: 0
          maximum: 100
        type_suivi_stock:
          type: string
          maxLength: 20
        quantite_min_commande:
          type: integer
        niveau_reappro_quantite:
          type: integer
        stock_minimum:
          type: integer
        stock_maximum:
          type: integer
          nullable: true
        activer_suivi_stock:
          type: boolean
        id_fournisseur_principal:
          type: integer
          nullable: true
        reference_fournisseur:
          type: string
          maxLength: 100
          nullable: true
        delai_livraison_fournisseur_jours:
          type: integer
          nullable: true
        compte_general_vente:
          type: string
          maxLength: 20
          nullable: true
        compte_general_achat:
          type: string
          maxLength: 20
          nullable: true
        code_taxe_vente:
          type: string
          maxLength: 10
          nullable: true
        code_taxe_achat:
          type: string
          maxLength: 10
          nullable: true
    ProductPaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties:
            type: object
            properties:
              message:
                type: string
              code:
                type: string
        timestamp:
          type: string
          format: date-time